ControlSettings to ts
